refactor(cost-explorer): move helper func from helper.ts to module file
