+++ src/services/cost-explorer/modules/CostExplorerFilterTags.vue
@@ -10,7 +10,7 @@
         <template v-else>
             <p-tag v-for="(item, idx) in tagItems" :key="`selected-tag-${idx}-${item.value}`"
                    :deletable="!printMode && deletable"
-                   :category-item="item.categoryItem"
+                   :category-item="hideCategory ? undefined : item.categoryItem"
                    :key-item="item.keyItem"
                    :value-item="item.valueItem"
                    @delete="handleDeleteFilterTag(item)"
@@ -32,7 +32,7 @@ import { cloneDeep } from 'lodash';
 
 import { store } from '@/store';
 
-import { getRefinedTagItems } from '@/services/cost-explorer/lib/helper';
+import { FILTER, FILTER_ITEM_MAP } from '@/services/cost-explorer/lib/config';
 import type { CostFiltersMap } from '@/services/cost-explorer/type';
 
 
@@ -63,6 +63,10 @@ export default defineComponent<Props>({
             type: Object,
             default: () => ({}),
         },
+        hideCategory: {
+            type: Boolean,
+            default: false,
+        },
         deletable: {
             type: Boolean,
             default: false,
@@ -79,6 +83,32 @@ export default defineComponent<Props>({
             })),
             tagItems: computed<TagItem[]>(() => getRefinedTagItems(state.resourceMap, props.filters)),
         });
+
+        /* Util */
+        const getRefinedTagItems = (resourceMap, filters: CostFiltersMap): TagItem[] => {
+            const results: TagItem[] = [];
+            Object.entries(filters).forEach(([category, filterItems]) => {
+                const resourceItems = resourceMap[category];
+                filterItems.forEach((item) => {
+                    let valueLabel = item.v;
+                    const resourceItem = resourceItems?.[item.v];
+                    if (resourceItem) {
+                        valueLabel = category === FILTER.REGION ? resourceItem?.name : resourceItem?.label;
+                    }
+                    let keyItem: KeyItem | undefined;
+                    if (category !== item.k) {
+                        const convertedTagKey = (item.k as string).replace(`${category}.`, '');
+                        keyItem = { name: convertedTagKey, label: convertedTagKey };
+                    }
+                    results.push({
+                        categoryItem: { name: category, label: FILTER_ITEM_MAP[category].label },
+                        keyItem,
+                        valueItem: { name: item.v, label: valueLabel },
+                    });
+                });
+            });
+            return results;
+        };
 
         /* Event */
         const handleDeleteFilterTag = (item: TagItem) => {
